Create MyLibrary
This is a project organizing my personal library and cataloging information on each book based on their titles, the name of the author, the number of pages in the book, the genre, ISBN identifier, and a variety of other information. With this data in the SQL database, I have the opportunity to explore patterns in my reading habits as well as a strategic method of keeping track of my book collection.