Camera Store
This is a project I created to practice inserting data into SQL. I have passion for photography and cameras, so I developed a hypothetical camera store and this database stores information on the cameras by their brand, file type, model, year developed, and price in the store. This is information critical to any small business owner to keep track of inventory and sales patterns.